# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "APIs",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List all Users",
        "description": "Returns a list of all users, optionally filtered by search term",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "A search term to filter the list of users"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/{id}/friends": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Users friends",
        "description": "Returns array of Users, that are friends with current user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of User whose friends requesting"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "User Detail",
        "description": "Returns extended user object",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "username of user account"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDetail"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/{id}/follow": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Follow album",
        "description": "Returns extended user object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of user"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "albumId": {
                  "type": "number",
                  "example": 111
                }
              },
              "required": [
                "albumId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDetail"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/{id}/like": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Like post",
        "description": "Returns extended user object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of user"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "postId": {
                  "type": "number",
                  "example": 111
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDetail"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/friends/remove": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Unfriend user",
        "description": "Returns extended user object",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "id",
            "in": "param",
            "required": true,
            "type": "string",
            "description": "ID of user"
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "friendId": {
                  "type": "string",
                  "example": "UUID"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "description": "Returns extended user object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "email"
                },
                "password": {
                  "type": "string",
                  "example": "new password"
                },
                "profileName": {
                  "type": "string",
                  "example": "new profile name"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDetail"
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "description": "Returns extended user object",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "152521515"
        },
        "username": {
          "type": "string",
          "example": "SomeName"
        },
        "profileName": {
          "type": "string",
          "example": "Dude 1"
        },
        "email": {
          "type": "string",
          "example": "thisIsTheemail@gmail.com"
        }
      }
    },
    "UserList": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/User"
      }
    },
    "UserDetail": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "152521515"
        },
        "username": {
          "type": "string",
          "example": "SomeName"
        },
        "profileName": {
          "type": "string",
          "example": "Dude 1"
        },
        "email": {
          "type": "string",
          "example": "thisIsTheemail@gmail.com"
        },
        "followedAlbums": {
          "$ref": "#/definitions/AlbumList"
        },
        "ownedAlbums": {
          "$ref": "#/definitions/AlbumList"
        },
        "friends": {
          "$ref": "#/definitions/UserList"
        },
        "friendedBy": {
          "$ref": "#/definitions/UserList"
        },
        "likedPost": {
          "$ref": "#/definitions/PostList"
        },
        "posts": {
          "$ref": "#/definitions/PostList"
        }
      }
    },
...
    "Album": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 111
        },
        "title": {
          "type": "string",
          "example": "Stretnutie na plazi"
        },
        "description": {
          "type": "string",
          "example": ""
        },
        "ownerId": {
          "type": "string",
          "example": "152521515"
        },
        "tags": {
          "type": "array",
          "example": [
            "leto",
            "more",
            "relax"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "AlbumList": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Album"
      }
    },
...
    "Post": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 15
        },
        "title": {
          "type": "string",
          "example": "Potapanie sa"
        },
        "photo": {
          "type": "object",
          "properties": {
            "0": {
              "type": "number",
              "example": 47
            },
            "1": {
              "type": "number",
              "example": 57
            },
            "2": {
              "type": "number",
              "example": 106
            },
            "3": {
              "type": "number",
              "example": 47
            },
            "4": {
              "type": "number",
              "example": 52
            },
            "5": {
              "type": "number",
              "example": 65
            },
            "6": {
              "type": "number",
              "example": 65
            },
            "7": {
              "type": "number",
              "example": 81
            },
            "8": {
              "type": "number",
              "example": 83
            },
            "9": {
              "type": "number",
              "example": 107
            },
            "10": {
              "type": "number",
              "example": 90
            },
            "11": {
              "type": "number",
              "example": 74
            },
            "12": {
              "type": "number",
              "example": 82
            },
            "13": {
              "type": "number",
              "example": 103
            },
            "14": {
              "type": "number",
              "example": 65
            },
            "15": {
              "type": "number",
              "example": 66
            },
            "16": {
              "type": "number",
              "example": 65
            },
            "17": {
              "type": "number",
              "example": 81
            },
            "18": {
              "type": "number",
              "example": 69
            },
            "19": {
              "type": "number",
              "example": 65
            },
            "20": {
              "type": "number",
              "example": 83
            },
            "21": {
              "type": "number",
              "example": 65
            },
            "22": {
              "type": "number",
              "example": 66
            },
            "23": {
              "type": "number",
              "example": 73
            },
            "24": {
              "type": "number",
              "example": 65
            },
            "25": {
              "type": "number",
              "example": 65
            },
            "26": {
              "type": "number",
              "example": 68
            },
            "27": {
              "type": "number",
              "example": 47
            }
          }
        },
        "description": {
          "type": "string",
          "example": "Nie vsetky dobre veci su napovrchu"
        },
        "userId": {
          "type": "string",
          "example": "152521515"
        },
        "albumId": {
          "type": "number",
          "example": 111
        },
        "locationId": {
          "type": "number",
          "example": 145
        }
      }
    },
...
    "PostList": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Post"
      }
    },
...
  }
}